eval: commonize getting service routes to fix the wrong knowledge graph of sockshop
